[BEAM-12963]: reorganize project structure
