+++ playground/frontend/lib/modules/output/components/output_area.dart
